+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 #Unreleased
 * Added `displayCustom`, which allows ability to render custom component in the centre of the knob, if `displayInput` is set to false.
-
+* Added prop to disable mouse wheel action.
 
 # 0.3.0
 * Added boolean property `disableTextInput`